+++ localflavor/tw/tw_choices.py
@@ -1,0 +1,28 @@
+#: An alphabetical list of cities and counties for use as `choices` in a formfield.
+#: Based on the administrative divisions of Taiwan.
+#: https://en.wikipedia.org/wiki/List_of_administrative_divisions_of_Taiwan
+
+TW_ADMINISTRATIVE_DIVISION_CHOICES = (
+    ("changhua_county", "彰化縣"),
+    ("chiayi_city", "嘉義市"),
+    ("chiayi_county", "嘉義縣"),
+    ("hsinchu_city", "新竹市"),
+    ("hsinchu_county", "新竹縣"),
+    ("hualien_county", "花蓮縣"),
+    ("kaohsiung_city", "高雄市"),
+    ("keelung_city", "基隆市"),
+    ("kinmen_county", "金門縣"),
+    ("lienchiang_county", "連江縣"),
+    ("miaoli_county", "苗栗縣"),
+    ("nantou_county", "南投縣"),
+    ("new_taipei_city", "新北市"),
+    ("penghu_county", "澎湖縣"),
+    ("pingtung_county", "屏東縣"),
+    ("taichung_city", "台中市"),
+    ("tainan_city", "台南市"),
+    ("taipei_city", "台北市"),
+    ("taitung_county", "台東縣"),
+    ("taoyuan_city", "桃園市"),
+    ("yilan_county", "宜蘭縣"),
+    ("yunlin_county", "雲林縣"),
+)